+++ web-app/src/app/app.effects.ts
@@ -1,6 +1,6 @@
 import { CoreEffects } from './core/store/core.effects';
 import { DataTypesEffects } from './core/datatypes/store/data-types.effects';
-import { EepDataEffects } from './eep/data/store/eep-data.effects';
+import { EepDataEffects } from './client/modules/data/store/eep-data.effects';
 import { GenericDataEffects } from './eep/generic-data/store/generic-data.effects';
 import { IntersectionEffects } from './eep/intersection/store/intersection.effects';
 import { SignalEffects } from './eep/signals/store/signal.effects';
