+++ web-app/src/app/client/modules/data/eep-data.module.ts
@@ -1,18 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SharedModule } from '../../shared/shared.module';
+import { SharedModule } from '../../../shared/shared.module';
 import { EepDataListComponent } from './data-list/eep-data-list.component';
 import { EepDataRoutingModule } from './data-routing.module';
 
 @NgModule({
-  declarations: [
-    EepDataListComponent,
-  ],
-  imports: [
-    CommonModule,
-    EepDataRoutingModule,
-    SharedModule,
-  ],
+  declarations: [EepDataListComponent],
+  imports: [CommonModule, EepDataRoutingModule, SharedModule],
 })
-export class EepDataModule {
-}
+export class EepDataModule {}
