+++ web-app/src/app/client/modules/data/models/eep-data.model.ts
